requirements.txt
boto3
fastapi
fire
humanize
mypy
pyyaml
